style: move the interface in the objc extension to the original header file
Since the spm could not load the interface of the objc extension we need
to move those interface back to the FLAnimatedImage.h header to let client
could successfully use it in Swift.
