+++ FLAnimatedImage/FLAnimatedImage+WebP.m
@@ -4,7 +4,6 @@
 //  Created by Grant Paul.
 //  Copyright (c) 2015 Facebook. All rights reserved.
 
-#import "FLAnimatedImage+WebP.h"
 #import "FLAnimatedImage+Internal.h"
 #import "FLAnimatedSystemWebPDataSource.h"
 #import "FLAnimatedWebPDataSource.h"
