--- FLAnimatedImage/include/FLAnimatedImage+GIF.h
@@ -1,16 +1,0 @@
-//
-//  FLAnimatedImage+GIF.h
-//  Facebook
-//
-//  Created by Grant Paul.
-//  Copyright (c) 2015 Facebook. All rights reserved.
-//
-
-#import "FLAnimatedImage.h"
-
-@interface FLAnimatedImage (GIF)
-
-+ (FLAnimatedImage *)animatedImageWithGIFData:(NSData *)data;
-
-@end
-
